% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch-aux-fcns.R
\name{center_out_uni}
\alias{center_out_uni}
\title{Create a univariate grid of values from the center out}
\usage{
center_out_uni(center, increment, length)
}
\arguments{
\item{center}{Output of \code{get_initial_beta_D} or a number to center
the grid (numeric)}

\item{increment}{Granularity of the grid (numeric)}

\item{length}{Number of grid values to the right and left of the grid
(numeric)}
}
\value{
A vector of grid values
}
\description{
Given some center, extend in the positive and negative directions to create
a grid of values
}
