% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_lee.R
\name{chen_lee}
\alias{chen_lee}
\title{Simulate data according to Chen and Lee (2017)}
\usage{
chen_lee(n = 500, p_D = 3)
}
\arguments{
\item{n}{Number of observations; defaults to 500 (numeric)}

\item{p_D}{Number of endogeneous variables; defaults to 3
(numeric)}
}
\value{
A named list:
\enumerate{
\item Y: outcome variable (n by 1 matrix)
\item D: endogeneous variable (n by p_D matrix)
\item Z: instruments (n by p_D matrix)
\item X: matrix of 1's (n by 1 matrix)
\item errors: matrix of errors and shocks (n by (p_D + 1) matrix); first
column is the vector of errors on the location scale model; all other
columns are shocks to the instruments when defining D.
}
}
\description{
Given the number of observations and number of endogeneous variables, create
an outcome variable defined by a location scale model where the coefficients
on the endogenous variables are supposed to be 1.
}
\details{
The error term in the location scale model that underpins this simulation
design is defined in terms of the endogeneous variables (which is why we
call these variables "endogenous"). To properly estimate the coefficients on
the endogeneous variables, we require instruments that are uncorrelated with
the errors, related to the endogeneous variables, and only related to the
outcome variable through their association with these endogeneous variables.

This function creates errors, endogeneous variables, instruments, and an
outcome variable such that the above terms are satisfied.
The errors are drawn independently of the instruments from a multivariate
normal distribution. The instruments are drawn from a standard normal
normal distribution. The endogeneous variables are multiples of the
cumulative distribution function of the shocked instruments.
The error in the true model for the outcome variable is defined in terms
of the endogeneous variables.

The original Chen and Lee simulation design used 3 endogeneous variables.
This design allows for an arbitrary number of endogeneous variables.
To allow fewer endogeneous variables, say 2 endogeneous variables, we simply
omit the third endogeneous variable from the original Chen and Lee
simulation before constructing our outcome variable.
}
