% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{concatenate_csvs}
\alias{concatenate_csvs}
\title{Combine directory of CSVs into a single CSV}
\usage{
concatenate_csvs(
  dir,
  cols,
  pattern = NULL,
  read = utils::read.csv,
  ...,
  remove_after_merge = FALSE
)
}
\arguments{
\item{dir}{Directory of CSVs}

\item{cols}{Vector of column names; if NULL (default), default column names
are applied to the merged CSV (character vector)}

\item{pattern}{Regular expression to filter files in \code{dir};
If NULL (default), all files in \code{dir} will be read}

\item{read}{Function to read CSV files; defaults to \code{read.csv} but
other options include \code{data.table::fread}}

\item{...}{Arguments to be passed to \code{read}}

\item{remove_after_merge}{If FALSE (default), files in \code{dir} are not
removed (boolean)}
}
\value{
A CSV called "merged.csv" in \code{dir} directory
}
\description{
Combine directory of CSVs into a single CSV
}
