% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{decompose_A}
\alias{decompose_A}
\title{Decompose constraint matrix}
\usage{
decompose_A(
  Y,
  X,
  D,
  Z,
  O_neg = NULL,
  O_pos = NULL,
  Phi = linear_projection(D, X, Z)
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{O_neg, O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}
}
\value{
A vector whose length is the same as the number of constraints;
\enumerate{
\item pf: primal feasibility
\item df_X: dual feasibility, X
\item df_Phi: dual feasibility, Phi
\item cs_uk: complementary slackness, u and k
\item cs_vl: complementary slackness, v and l
\item cs_ak: complementary slackness, a and k
\item cs_al: complementary slackness, a and l
\item pp_a: preprocessing, a
\item pp_k: preprocessing, k
\item pp_l: preprocessing, l
}
}
\description{
Print what each row of the A matrix represents
}
\details{
Using the dimensions of the inputs, this function prints the constraints
associated with each row.
}
