% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{density_wald}
\alias{density_wald}
\title{Evaluate Asymptotic Distribution of IQR Estimator}
\usage{
density_wald(beta_star, beta_opt, varcov_mat)
}
\arguments{
\item{beta_star}{Vector of proposed coefficients beta_D and beta_X
(vector of length p_D + p_X)}

\item{beta_opt}{Vector of beta_D and beta_X coefficients from IQR point
estimate (vector of length p_D + p_X)}

\item{varcov_mat}{Asymptotic variance-covariance matrix; see
\code{wald_varcov} and \code{naive_varcov}}
}
\value{
density of limiting distribution of estimator evaluated at
\code{beta_star}
}
\description{
Evaluate Asymptotic Distribution of IQR Estimator
}
