% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{Diagnostic}
\alias{Diagnostic}
\title{Plotting the Objective Function}
\usage{
Diagnostic(object, i, size = 0.05, trim = NULL, GMM_only = 0)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}

\item{i}{index of specific quantile to inspect. For example, if
argument \code{taus = c(0.25,0.5,0.75)} was used as an input of \code{ivqr()}
, specify \code{i=2} for the median.}

\item{size}{the size of the test. Default is 0.05.}

\item{trim}{a vector of two number indicating the lower and upper bounds 
of the quantiles to inspect.}
}
\value{
A plot of the objective function as well as the (asymptotical) first
order conditions evaluated at the coefficient estimates.
}
\description{
Plotting the Objective Function
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, c(0.25,0.5,0.75), grid = seq(-2,2,0.2), data = ivqr_eg)
Diagnostic(fit,2) # Plot the objective function at the median.
}
