% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{draw_proposal_h}
\alias{draw_proposal_h}
\title{Propose active basis}
\usage{
draw_proposal_h(
  weights,
  p,
  num_draws,
  label_bool = FALSE,
  label_skip = 5,
  existing_h = c()
)
}
\arguments{
\item{weights}{Vector of weights for each observation}

\item{p}{Dimension of design matrix in IQR-QR (p_X + p_Phi)}

\item{num_draws}{Number of draws from proposal distribution of active basis}

\item{theta}{Tuning parameter}
}
\value{
indices in the active basis
}
\description{
Indices whose observations have small residuals will be more likely to
belong to the active basis, while indices whose observations have extreme
residuals will be less likely to belong to the active basis.
}
