% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{education_autor}
\alias{education_autor}
\alias{Y_educ}
\alias{X_educ}
\alias{D_educ}
\alias{Z_educ}
\alias{EducFE_educ}
\alias{DInt_educ}
\alias{ZInt_educ}
\alias{y_name}
\alias{d_name}
\alias{z_name}
\alias{covs_name}
\alias{educfe_name}
\alias{dint_name}
\alias{zint_name}
\title{Education-relevant Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 1 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 50 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 1 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 1 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 24684 rows and 2 columns.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 50.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.
}
\usage{
Y_educ

X_educ

D_educ

Z_educ

EducFE_educ

DInt_educ

ZInt_educ

y_name

d_name

z_name

covs_name

educfe_name

dint_name

zint_name
}
\description{
24684 observations with known education status
\describe{
\item{\code{Y_educ}}{outcome variable, column matrix}
\item{\code{X_educ}}{covariate matrix, 50-column matrix}
\item{\code{D_educ}}{endogeneous matrix, column matrix (emp)}
\item{\code{Z_educ}}{instrument matrix, column matrix (emp_pct)}
\item{\code{EducFE_educ}}{education fixed effects matrix, 2-column matrix (hsgrad, hsplus)}
\item{\code{DInt_educ}}{endogeneous x education matrix, 2-column matrix}
\item{\code{ZInt_educ}}{instrument x education matrix, 2-column matrix}
}
}
\keyword{datasets}
