% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat-fcns.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Expand a Matrix}
\usage{
expand_matrix(
  mat,
  newrow,
  newcol,
  row_direction,
  col_direction,
  fill = 0,
  quietly = TRUE
)
}
\arguments{
\item{mat}{An m-by-n matrix to embed into a larger p-by-q matrix}

\item{newrow}{The number of rows of the larger matrix (p)}

\item{newcol}{The number of columns of the larger matrix (c)}

\item{row_direction}{Adds rows to "top" or "bottom" of \code{mat}}

\item{col_direction}{Adds columns to "left" or "right" of \code{mat}}

\item{fill}{The value of entries in the new rows and columns; defaults to 0}

\item{quietly}{If TRUE, suppresses messages; defaults to TRUE}
}
\value{
a larger matrix that embeds \code{mat}
}
\description{
Expand an m-by-n matrix to be a larger p-by-q matrix
}
\details{
This function is useful to write a simple matrix whose entries can themselves
be grouped as a matrix.
}
\examples{
\dontrun{
expand_matrix(matrix(c(1, 2, 3, 4), nrow = 2, byrow = TRUE),
              newrow = 3, newcol = 3,
              row_direction = "bottom",
              col_direction = "right")
}

}
\seealso{
\code{\link{block_diagonal}}

Other matrix operations: 
\code{\link{block_diagonal}()}
}
\concept{matrix operations}
