% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foc-center-archive.R
\name{find_subsample_in_polytope}
\alias{find_subsample_in_polytope}
\title{Find subsample in the FOC polytope}
\usage{
find_subsample_in_polytope(
  h,
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  beta_D_proposal = NULL,
  beta_X_proposal = NULL,
  subsample_size,
  params = list(OutputFlag = 0),
  type = "C"
)
}
\arguments{
\item{h}{Active basis in terms of the data provided}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (numeric)}

\item{beta_D_proposal}{Coefficients on the endogeneous variables (vector of
length p_D); if NULL, use \code{h_to_beta} function and the \code{h}
argument to determine \code{beta_D_proposal}}

\item{beta_X_proposal}{Coefficients on the exogeneous variables (vector of
length p_D); if NULL, use \code{h_to_beta} function and the \code{h}
argument to determine \code{beta_X_proposal}}

\item{subsample_size}{Size of subsample}

\item{params}{Named list of parameters to send to Gurobi}

\item{type}{Either "C" or "B" to denote if our "omega" variables are
continuous or binary; If "C", then the solution is the continuous center
of our FOC polytope; If "B", then the solution is the actual center}
}
\description{
It is useful to start our random walk at a subsample that is inside the FOC
polytope. To find this polytope, we solve a linear program that aims to
minimize the absolute L-1 norm of the weighted average of the xi_i objects.
The weights are inside the unit interval and sum to the size of our
subsample. Most of the weights will be integral, but some will be between 0
and 1. We will round the non-integral solutions so they are integra. If the
weight for an observaion is 1, then that observation belongs in the
subsample. Otherwise, the observation doesn't belong in the subsample. Due
to the rounding, it isn't guaranteed that the resulting subsample belongs in
the FOC polytope, but it shouldn't be too far away from the polytope.
Note that this program excludes the observations in the active basis -- we
must add them back in after-the-fact.
}
