% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-subsample-archive.R
\name{first_approach_v4}
\alias{first_approach_v4}
\title{Propose subsamples}
\usage{
first_approach_v4(
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  h,
  subsample_size,
  beta_D_proposal = NULL,
  beta_X_proposal = NULL,
  gamma = 1,
  l_norm = 1,
  l_power = l_norm
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile (numeric)}

\item{h}{Indices of active basis (vector of length p_X + p_Phi)}

\item{subsample_size}{Size of subsample (numeric at most n)}

\item{beta_D_proposal}{Coefficients on the endogeneous variables (vector of
length p_D); if NULL, use \code{h_to_beta} function and the \code{h}
argument to determine \code{beta_D_proposal}}

\item{beta_X_proposal}{Coefficients on the exogeneous variables (vector of
length p_D); if NULL, use \code{h_to_beta} function and the \code{h}
argument to determine \code{beta_X_proposal}}

\item{gamma, l_norm, l_power}{Hyperparameters}
}
\value{
Named list
\enumerate{
\item \code{subsample_set}: set of indices in subsample
\item \code{subsample_weights}: set of weights for each observation in subsample
\item \code{prob}: probability of proposing \code{subsample_set} (except
for observations already in active basis)
\item \code{log_prob}: log of \code{prob}
\item \code{xi}: xi vector for this subsample
}
}
\description{
Propose observations, one at a time, until we have enough to create a subsample.
Note that observations inside the active basis will be inside the final subsample.
Unlike the original \code{first_approach}, I am replacing the exponential
function with the reciprocal.
}
