% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_subsampling.R
\name{foc_membership}
\alias{foc_membership}
\title{Verify membership of a data set in the FOC conditions}
\usage{
foc_membership(
  h,
  Y_subsample,
  X_subsample,
  D_subsample,
  Phi_subsample,
  tau,
  beta_D = h_to_beta(h, Y = Y_subsample, X = X_subsample, D = D_subsample, Phi =
    Phi_subsample)$beta_D
)
}
\arguments{
\item{h}{Indices of the active basis written in terms of the subsample data
\link{p-dimensional vector}}

\item{Y_subsample}{Outcome vector in the subsample \link{m by 1 matrix}}

\item{X_subsample}{Covariates in subsample \link{m by p_X matrix}}

\item{D_subsample}{Endogeneous variables in subsample \link{m by p_D matrix}}

\item{Phi_subsample}{Transformed instruments in subsample \link{m by p_Phi}}

\item{tau}{Quantile \link{numeric}}

\item{beta_D}{Coefficients on the endogeneous variable; ideally obtained
from \code{h} \link{p_D by 1 matrix}}
}
\value{
TRUE if the subsample satisfies FOC conditions; FALSE otherwise
}
\description{
From the active basis, we derive the FOC conditions. If a data set satisfies
the FOC conditions, then we know that the coefficients obtained from this
active basis solve the IQR problem for this data set.
}
\seealso{
Other mcmc_subsampling: 
\code{\link{h_to_beta}()}
}
\concept{mcmc_subsampling}
