% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foc_membership.R
\name{foc_membership_v2}
\alias{foc_membership_v2}
\title{Verify membership of a data set in the FOC conditions}
\usage{
foc_membership_v2(
  h,
  Y_subsample,
  X_subsample,
  D_subsample,
  Phi_subsample,
  tau,
  beta_D = h_to_beta(h, Y = Y_subsample, X = X_subsample, D = D_subsample, Phi =
    Phi_subsample)$beta_D,
  tolerance = 1e-09
)
}
\arguments{
\item{h}{Indices of the active basis written in terms of the subsample data
(p-dimensional vector)}

\item{Y_subsample}{Outcome vector in the subsample (m by 1 matrix)}

\item{X_subsample}{Covariates in subsample (m by p_X matrix)}

\item{D_subsample}{Endogeneous variables in subsample (m by p_D matrix)}

\item{Phi_subsample}{Transformed instruments in subsample (m by p_Phi)}

\item{tau}{Quantile (numeric)}

\item{beta_D}{Coefficients on the endogeneous variable; ideally obtained
from \code{h} (p_D by 1 matrix)}
}
\value{
Named list
\enumerate{
\item \code{status}: TRUE if subsample satisfies FOC conditions; FALSE
otherwise
\item \code{beta_Phi}: coefficients on transfomed instruments when
running a QR of Y - D \%*\% beta_D on X and Phi, where beta_D is given by
\code{h}
\item \code{norm}: L1 norm of \code{beta_Phi}
}
}
\description{
If a data set satisfies the FOC conditions with respect to some active
basis, then a quantile regression of the concentrated-out outcome using the
endogeneous coefficients given by the active basis on the covariates and the
transformed instruments should be 0.
}
\seealso{
Other mcmc_subsampling: 
\code{\link{foc_membership}()},
\code{\link{h_to_beta}()}
}
\concept{mcmc_subsampling}
