% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{get_initial_beta_D}
\alias{get_initial_beta_D}
\title{Obtain naive coefficient estimates for \code{beta_D}}
\usage{
get_initial_beta_D(Y, X, D, Z, tau, ...)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile of interest (numeric between 0 and 1)}

\item{...}{Arguments to be passed to \code{quantreg::rq()}}
}
\value{
A vector of coefficients on the endogenous variable
}
\description{
Naively regress \code{Y} on \code{D}, \code{Z}, and \code{X}, and return
the coefficients on \code{D}.
}
