% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{get_iqr_objective}
\alias{get_iqr_objective}
\title{Get value of IQR objective given coefficients on endogeneous variable}
\usage{
get_iqr_objective(beta_D, Y, X, D, Z, tau, ...)
}
\arguments{
\item{beta_D}{Vector of coefficients on the endogenous variable
(numeric vector)}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile of interest (numeric between 0 and 1)}

\item{...}{Arguments to be passed to \code{quantreg::rq()}}
}
\value{
A named list of two entries:
\enumerate{
\item \code{beta_Z}: named vector of coefficients on the instruments
\item \code{tau}: quantile of interest
\item \code{obj}: sum of absolute value of \code{beta_Z}, i.e., value of
IQR objective
}
}
\description{
Obtain the sum of the absolute value of the coefficients on the instruments
in a quantile regression of \code{Y} after concentrating out \code{D}
according to \code{beta_D} on \code{Z} and \code{X}.
}
