% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{get_iqr_objective_grid}
\alias{get_iqr_objective_grid}
\title{Compute IQR objective given grid of coefficients on endogeneous variables}
\usage{
get_iqr_objective_grid(
  grid,
  Y,
  X,
  D,
  Z,
  tau,
  update = round(nrow(grid)/20),
  ...
)
}
\arguments{
\item{grid}{Data frame with p_D columns where each row is a set of
\code{beta_D} coefficients; Output of \code{center_out_grid}}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile of interest (numeric between 0 and 1)}

\item{update}{Send progress after every \code{update} values of grid search;
defaults ot \code{round(nrow(grid)) / 20}, i.e., 20 updates (numeric)}

\item{...}{Arguments to be passed to \code{quantreg::rq()}}
}
\value{
A data frame of dimension \code{nrow(grid)} by p_D + p_Z + 1 where
each row corresponds to one set of coordinate values on the grid, the
corresponding values for the instrument coefficient, and the resulting
IQR objective
}
\description{
For each set of \code{beta_D} suggested by \code{grid}, compute the
sum of the absolute values of \code{beta_Z}
}
\details{
This code is not run in parallel.
}
