% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{get_iqr_objective_grid_parallel}
\alias{get_iqr_objective_grid_parallel}
\title{Compute IQR objective given grid of coefficients on endogeneous variables}
\usage{
get_iqr_objective_grid_parallel(
  grid,
  Y,
  X,
  D,
  Z,
  tau,
  log_dir = NULL,
  cores = parallel::detectCores()[1] - 2,
  ...
)
}
\arguments{
\item{grid}{Data frame with p_D columns where each row is a set of
\code{beta_D} coefficients; Output of \code{center_out_grid}}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile of interest (numeric between 0 and 1)}

\item{log_dir}{Path of log directory to store parallelized results;
if NULL (default), log is not saved (string)}

\item{cores}{Number of cores to be used in parallelization process}

\item{...}{Arguments to be passed to \code{quantreg::rq()}}
}
\value{
A data frame of dimension \code{nrow(grid)} by p_D + p_Z + 1 where
each row corresponds to one set of coordinate values on the grid, the
corresponding values for the instrument coefficient, and the resulting
IQR objective
}
\description{
For each set of \code{beta_D} suggested by \code{grid}, compute the
sum of the absolute values of \code{beta_Z}
}
\details{
This code is run in parallel.
To store results as they are found, specify a directory with \code{log_dir}
to store intermediate CSV files. Then, use \code{concatenate_csvs} to
merge CSVs into one file, i.e.,
\code{cols = c("iteration", colnames(grid), colnames(Z), "objective")}
\code{concatenate_csv(dir = log_dir, cols = cols, header = FALSE)}

Note: if the function terminates without any problems, the intermediary
CSV files will be deleted, and the results are saved as a CSV in
\code{log_dir} as "gridsearch_results.csv".
}
