% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{h_to_beta}
\alias{h_to_beta}
\title{Find the coefficients given the active basis}
\usage{
h_to_beta(h, Y, X, D, Z, Phi = linear_projection(D, X, Z))
}
\arguments{
\item{h}{Active basis in terms of the data provided}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}
}
\value{
A named list of coefficients
\enumerate{
\item \code{beta_D}: coefficients on the endogeneous variables
\item \code{beta_X}: coefficients on the exogenous variables
\item \code{beta_Phi}: coefficients on the transformed instruments
}
}
\description{
Given the active basis, find the coefficients that solve the IQR problem.
}
\seealso{
Other mcmc_subsampling: 
\code{\link{foc_membership_v2}()},
\code{\link{foc_membership}()}
}
\concept{mcmc_subsampling}
