% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_milp.R
\name{iqr_milp}
\alias{iqr_milp}
\title{Compute inverse quantile regression estimator}
\usage{
iqr_milp(
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  solver = c("gurobi", "highs", "ecos"),
  ...
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{solver}{Choice of solver: "gurobi" (default), "highs", or "ecos"}

\item{...}{Arguments passed to \code{iqr_milp_gurobi}, \code{iqr_milp_highs}, or \code{iqr_milp_ecos}}
}
\description{
Solve a mixed integer linear program to compute the inverse
quantile regression estimator using either Gurobi or HiGHS.
}
