% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ivqr.ks}
\alias{ivqr.ks}
\title{General Inference
Testing general hypothesis regarding the qunatile process of the endogenous
variable.}
\usage{
ivqr.ks(object, variable = NULL, trim = c(0.05, 0.95), B = 2000,
  size = c(0.1, 0.05, 0.01, 0.001), nullH = "No_Effect", ...)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}

\item{variable}{A number indicates which endogenous variable to test. Since 
at most two endongenous variables can be included in the function \code{ivqr},
this argument should be either 1 or 2.}

\item{trim}{A vector of two numbers indicating the lower and upper bounds 
of the quantiles to consider.}

\item{B}{Number of sub-sampling in the bootstrap. Default is 2000.}

\item{size}{A vector indicates the desired size of the test. Critical values
will be reported accordingly.}

\item{nullH}{The hypothesis to be tested. The four options are: No_Effect, 
Dominance, Location_Shift, and Exogeneity.}
}
\value{
An ivqr_ks object which contains information regarding test statistics,
critical value, sub-sampling block size, ...etc.
}
\description{
General Inference
Testing general hypothesis regarding the qunatile process of the endogenous
variable.
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.15,0.85,0.02), grid = seq(-2,2,0.2), data = ivqr_eg) # taus should be a fine grid 
ivqr(fit,nullH=No_Effect) # Test of no effect.
ivqr(fit,nullH=Dominance) # Test of dominance.
ivqr(fit,nullH=Location_Shift) # Test of location shift.
ivqr(fit,nullH=Exogeneity) # Test of exogeneity.
}
