% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_confint.R
\name{line_confint}
\alias{line_confint}
\title{Construct a univariate confidence interval via a line search}
\usage{
line_confint(
  index,
  endogeneous = TRUE,
  beta_null = NULL,
  stopping_tolerance = NULL,
  width_ratio = 1,
  step_size = NULL,
  step_rate = 0.5,
  cores = 2,
  log_dir = NULL,
  log_name = "line_search.csv",
  remove_intermediate_csvs = FALSE,
  return_setup = FALSE,
  p_val_tol = 1e-06,
  small_change_count_tol = 10,
  alpha = 0.1,
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  B = NULL,
  orthogonalize_statistic = FALSE,
  homoskedasticity = FALSE,
  kernel = "Powell",
  show_progress = TRUE,
  FUN = preprocess_iqr_milp,
  ...
)
}
\arguments{
\item{index}{Index of the coefficient of interest
(numeric between 1 and p_D if \code{endogeneous} is TRUE;
numeric between 1 and p_X if \code{endogeneous} is FALSE)}

\item{endogeneous}{If TRUE (default), \code{index} refers to the
coefficients on the endogeneous variables; if FALSE, \code{index} refers to
the coefficients on the exogeneous variables (boolean)}

\item{beta_null}{Initial value of line search; suggested value is given by
\code{iqr_milp}; if NULL (default), naive quantile regression point
estimate will be used (numeric)}

\item{stopping_tolerance}{Smallest step size before stopping the line
search; if NULL (default), takes on the value of \code{width_ratio} times
the width of the naive confidence interval (numeric)}

\item{width_ratio}{Determines the relationship between the size of naive
confidence interval and default value of \code{stopping_tolerance};}

\item{step_size}{Initial step size of line search; if NULL(default), use
half the width of the naive confidence interval (numeric)}

\item{step_rate}{Rate at which we decrease the step size when we cross the
border of the confidence interval; defaults to 0.5 (numeric, less than 1)
defaults to 1 (numeric)}

\item{cores}{Number of cores to be used in parallelization process; defaults
to 2 (no more than 2 needed)}

\item{log_dir}{Path of log directory to store parallelized results;
if NULL (default), log is not saved (string)}

\item{log_name}{Name of CSV file (including extension) to store results;
defaults to "gridsearch_results.csv" (string)}

\item{remove_intermediate_csvs}{If TRUE, intermediate csvs created during
while loop will be removed after line search is finished; defaults to FALSE}

\item{return_setup}{If TRUE, return the step size and naive quantile
regression summary without carrying out the line search; defaults to FALSE}

\item{p_val_tol, small_change_count_tol}{If the change in p-value is less
than \code{p_val_tol} for \code{small_change_count_tol} consecutive
iterations, we stop the while loop; defaults to 1e-6 and 10}

\item{alpha}{Alpha level of the test; defaults to 0.1; only used when
\code{homoskedasticity} is FALSE (numeric between 0 and 1)}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{B}{Matrix that enters numerator of test statistic
(n by |J| + |K| matrix); If NULL (default), this matrix is (Phi_J, X_K),
where Phi_J and X_K are the columns of Phi and X with indices J and K;}

\item{orthogonalize_statistic}{If TRUE, \eqn{\tilde{B}} will be used in
numerator of test statistic; defaults to FALSE; for advanced users only}

\item{homoskedasticity}{If TRUE, assume density of error at 0 is constant;
defaults to FALSE (boolean)}

\item{kernel}{Only active if \code{homoskedasticity} is FALSE; either
"Powell" (default) to use the Powell estimator or
"Gaussian" to use a Gaussian kernel; only used when
\code{homoskedasticity} is FALSE}

\item{show_progress}{If TRUE (default), sends progress messages during
execution (boolean); also passed to \code{FUN}}

\item{FUN}{Either \code{preprocess_iqr_milp} (default) or \code{iqr_milp}}

\item{...}{Arguments passed to \code{FUN}}
}
\description{
Search for the smallest/largest value of a particular coefficient where the
null hypothesis of the coefficient being equal to that value is not
rejected.
}
\details{
The suggested value of \code{beta_null} is given by the output of
\code{iqr_milp} or \code{preprocess_iqr_milp}.  For example, if \code{index}
is 2, \code{endogeneous} is TRUE, and the \code{iqr_milp} estimates are
stored in \code{iqr}, then \code{iqr$beta_D[2]} should be the value of
\code{beta_null}.
}
