% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{linear_projection}
\alias{linear_projection}
\title{Linearly project a vector onto the space spanned by other vectors}
\usage{
linear_projection(Y, ...)
}
\arguments{
\item{Y}{Vector or matrix that will be projected (n by p_Y matrix)}

\item{...}{"Regressors" whose span is the destination of the projection
(matrix of n rows)}
}
\value{
Linear projection of \code{Y} on \code{...}
(matrix of dimension nrow(Y) by ncol(Y))
}
\description{
Find the fitted values from a regression of each column of \code{Y} on
regressors, which are given by matrices in \code{...}.
}
\details{
Ensure that the rows of \code{Y} and \code{...} are the same!
}
