% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{manipulate_qr_residuals}
\alias{manipulate_qr_residuals}
\title{Manipulate residuals from a quantile regression}
\usage{
manipulate_qr_residuals(string_formula, tau, factor = 1, FUN = stats::sd, ...)
}
\arguments{
\item{string_formula}{A formula to be passed to \code{quantreg::rq} in the
the form of a string (string)}

\item{tau}{Quantile of interest (number between 0 and 1)}

\item{factor}{Number to be multiplied to the output of \code{FUN};
defaults to 1 (numeric)}

\item{FUN}{A function to manipulate the residuals;
defaults to \code{stats::sd}}

\item{...}{Arguments passed to \code{FUN}}
}
\value{
Result of \code{factor} multiplied by the output of \code{FUN}
evaluated at the quantile regression specified by \code{string_formula} and
\code{tau}
}
\description{
Apply a function to the residuals of a quantile regression
}
\details{
The purpose of this function is to define the big M constant in
\code{iqr_milp}, \code{miqcp_proj}, and potentially other
functions.

The formula relies on variables defined outside the scope of the function.
}
