% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{naive_varcov}
\alias{naive_varcov}
\title{Compute variance-covariance matrix of naive endogeneous QR}
\usage{
naive_varcov(Y, D, X, tau, psi = 1)
}
\arguments{
\item{psi}{Coefficient on variance-covariance matrix}
}
\value{
variance-covariance matrix of beta_D and beta_X coefficients when
running a naive endogeneous QR of Y on D and X without an intercept
}
\description{
Compute variance-covariance matrix of naive endogeneous QR
}
