% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-analysis.R
\name{onestep}
\alias{onestep}
\title{Check membership of one-step neighbors}
\usage{
onestep(
  subsample,
  reference,
  h,
  Y,
  X,
  D,
  Phi,
  tau,
  MEMBERSHIP_FCN = foc_membership_v3,
  ...
)
}
\arguments{
\item{subsample}{An n-long vector with m ones}

\item{reference}{An n-long vector from which we measure the distance to the
one-step neighbors}

\item{h}{A vector of indices that are the active basis}

\item{Y, X, D, Phi}{data}

\item{tau}{Quantile}

\item{MEMBERSHIP_FCN}{function for checking membership}

\item{...}{arguments for STATUS}
}
\value{
A named list with two elements
\enumerate{
\item distance: vector with distances between neighbors and \code{reference}
\item status: vector indicating whether neighbor is inside polytope
}
}
\description{
Check membership of one-step neighbors
}
