% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{p_val_interpolation}
\alias{p_val_interpolation}
\title{Perform P-value-weighted Linear Interpolation}
\usage{
p_val_interpolation(old_p_val, new_p_val, old_beta, new_beta, alpha)
}
\arguments{
\item{old_p_val, new_p_val}{Values of the p-value associated with
\code{old_beta} and \code{new_beta} respectively}

\item{old_beta, new_beta}{Values of the coefficient that are inside and
outside of the confidence interval (both can't be inside the CI nor can both
be outside)}

\item{alpha}{Level of the hypothesis test}
}
\value{
Named list:
\enumerate{
\item beta_border: New beta value that is the result of interpolation
\item pi: p-value-based weights
}
}
\description{
In line searches, we should interpolate between the two closest values that
bound the true boundary of the confidence interval according to the p-value.
}
\details{
When performing a line search in a given direction, the final two beta
values will surround the true beta value that is the bound of the confidence
interval. Interpolating between these two beta values will get us closer to
the true beta value. The weights of this interpolation is given by the
p-values of the final two beta values from the line search as well as the
p-value of the true beta value, which is the alpha-level.
}
\seealso{
\code{\link{line_confint}}, \code{\link{line_confint_interpolation}}
}
