% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{parse_mult_logs}
\alias{parse_mult_logs}
\title{Parse multiple log files}
\usage{
parse_mult_logs(
  log_dir,
  expr = "log$",
  information = "incumbent",
  value = "list"
)
}
\arguments{
\item{log_dir}{Directory of log files}

\item{expr}{Expression to limit files in \code{log_dir}; defaults to
searching for "log" extension}

\item{information}{Set to "incumbent"}

\item{value}{How should the information be returned? defaults to "list",
also can be "data.frame"}
}
\description{
Use \code{parse_single_log} across multiple log files.
Only valid for \code{information = "incumbent"}.
}
