% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{parse_single_log}
\alias{parse_single_log}
\title{Parse information from a Gurobi log file}
\usage{
parse_single_log(log_path, information = "incumbent")
}
\arguments{
\item{log_path}{Path to a Gurobi log file}

\item{information}{Vector with values:
\enumerate{
\item "incumbent"
}}
}
\description{
Obtain information on incumbent solutions, Gurobi version, etc. from a
Gurobi log file.
}
\details{
When \code{information} is "incumbent", this function will retrieve the
times at which Gurobi has found an incumbent solution either through
heuristics or through solving the problem at the node.
}
