% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-fcns.R
\name{parse_single_sol}
\alias{parse_single_sol}
\title{Parse a single solution (.sol) file}
\usage{
parse_single_sol(sol_path)
}
\arguments{
\item{sol_path}{Path to .sol file}
}
\description{
Obtain solution vector from a single .sol file.
}
