% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{plot.ivqr}
\alias{plot.ivqr}
\title{Visualizing the quantile process of the endogenous variable}
\usage{
\method{plot}{ivqr}(object, variable = 1, size = 0.05, trim = c(0.05,
  0.95), ...)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}

\item{variable}{A number indicates which endogenous variable to test. Since 
at most two endongenous variables can be included in the function \code{ivqr},
this argument should be either 1 or 2.}

\item{size}{A number indicating the desired size of the point-wise 
confident inverval.}

\item{trim}{A vector which representss an inverval. 
Extreme quantile(s) outside the interval will not be shown in the plot.}
}
\description{
Visualizing the quantile process of the endogenous variable
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.1,0.9,0.1), grid = seq(-2,2,0.2), data = ivqr_eg) # a process
plot(fit)
}
