% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{plot.ivqr_weakIV}
\alias{plot.ivqr_weakIV}
\title{Visualizing the Weak-IV Robust Confidence Region}
\usage{
\method{plot}{ivqr_weakIV}(object, ...)
}
\arguments{
\item{object}{An ivqr_weakIV object returned from the function \code{weakIVtest()}}
}
\description{
Visualizing the Weak-IV Robust Confidence Region
}
\details{
This function will be called when evaluating \code{weakIVtest()}. 
The case of two endogenous varaibles is not supported.
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.1,0.9,0.1), grid = seq(-2,2,0.2), data = ivqr_eg) # a process
plot(weakIVtest(fit))
}
