% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_iqr_milp_ecos.R
\name{preprocess_iqr_milp_ecos}
\alias{preprocess_iqr_milp_ecos}
\title{Compute IQR estimator by preprocessing MILP using ECOS}
\usage{
preprocess_iqr_milp_ecos(
  Y,
  D,
  X,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  weights = NULL,
  M = NULL,
  TimeLimit = NULL,
  globalTimeLimit = Inf,
  prop_alpha_initial = 0.7,
  r = 1.25,
  show_iterations = FALSE,
  quietly = TRUE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{...}{Arguments passed to \code{iqr_milp_ecos}}
}
\description{
Compute IQR estimator by preprocessing MILP using ECOS
}
