% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_iqr_milp.R
\name{preprocess_iqr_milp_gurobi}
\alias{preprocess_iqr_milp_gurobi}
\title{Compute IQR estimator by preprocessing MILP using Gurobi}
\usage{
preprocess_iqr_milp_gurobi(
  Y,
  D,
  X,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  weights = NULL,
  M = NULL,
  TimeLimit = NULL,
  globalTimeLimit = Inf,
  prop_alpha_initial = 0.7,
  r = 1.25,
  show_iterations = FALSE,
  LogFileExt = ".log",
  ...
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 2 times the largest absolute residual from
quantile regression of Y on X and D}

\item{TimeLimit}{Maximum time (in seconds) spent on a linear program;
defaults to heuristic (numeric)}

\item{globalTimeLimit}{Maximum time (in seconds) spent on the entire
preprocessing function; defaults to Inf (i.e., no time limit) (numeric)
thus, 1 - \code{prop_alpha_initial} is the proportion of observations whose
dual variables (i.e., sign of the residuals) are fixed at the start
(number between 0 and 1)}

\item{prop_alpha_initial}{Initial value of the bandwidth \eqn{alpha};}

\item{r}{Rate at which we increase the bandwidth (\eqn{alpha}) at every
iteration (number greater than 1)}

\item{show_iterations}{If TRUE, print the iteration number to the console;
defaults to FALSE (boolean)}

\item{LogFileExt}{Extension of Gurobi log file; If \code{LogFileName} is
empty, then Gurobi log won't be saved and this argument will be ignored;
defaults to "log" (string)}

\item{...}{Arguments that will be passed to \code{iqr_milp_gurobi}}
}
\description{
Compute IQR estimator by preprocessing MILP using Gurobi
}
