% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{print.ivqr}
\alias{print.ivqr}
\title{Printing the model fit from \code{ivqr()}}
\usage{
\method{print}{ivqr}(x, ...)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}
}
\description{
Printing the model fit from \code{ivqr()}
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.1,0.9,0.1), grid = seq(-2,2,0.2), data = ivqr_eg) # a process
fit
}
