% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{print.ivqr_weakIV}
\alias{print.ivqr_weakIV}
\title{Printing the Weak-IV Robust Confidence Region}
\usage{
\method{print}{ivqr_weakIV}(x, ...)
}
\arguments{
\item{x}{An ivqr_weakIV object returned from the function \code{weakIVtest()}}
}
\description{
Printing the Weak-IV Robust Confidence Region
}
\details{
The case of two endogenous varaibles is not supported.
}
