% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{proposal_h}
\alias{proposal_h}
\title{Compute unnormalized proposal density of active basis}
\usage{
proposal_h(weights, h)
}
\arguments{
\item{weights}{Weights on each index, see \code{weight_indices}}

\item{h}{Indices in the active basis}
}
\value{
unnormalized proposal density of active basis
}
\description{
Compute unnormalized proposal density of active basis
}
