% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{send_note_if}
\alias{send_note_if}
\title{Conditionally send note to console}
\usage{
send_note_if(note, condition, fcn, ...)
}
\arguments{
\item{note}{A string that will be sent to the console}

\item{condition}{If TRUE, message will be sent to console.}

\item{fcn}{A function, e.g., \code{\link[base]{message}}
or \code{\link[base]{warning}}}

\item{...}{Additional arguments for \code{fcn}}
}
\value{
A note in the console if \code{quietly} is FALSE
}
\description{
Send a message or warning if \code{condition} is TRUE
}
\details{
The intended values of \code{fcn} are \code{\link[base]{message}} and
\code{\link[base]{warning}}.
If \code{condition} is TRUE, then send the message to the console according
to \code{fcn}.
Otherwise, do not send anything to the console.
}
\examples{
\dontrun{
send_note_if("This is a warning.", TRUE, warning)
}

}
\seealso{
Other dev: 
\code{\link{collect_note_if}()}
}
\concept{dev}
