% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{summary.ivqr}
\alias{summary.ivqr}
\title{Summarizing the model fit returned from ivqr()}
\usage{
\method{summary}{ivqr}(x, i = NULL, ...)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}
}
\description{
Summarizing the model fit returned from ivqr()
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.1,0.9,0.1), grid = seq(-2,2,0.2), data = ivqr_eg) # a process
summary(fit)
}
