% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{target_h}
\alias{target_h}
\title{Evaluate target density in the MCMC Sampler of the proposal coefficients}
\usage{
target_h(...)
}
\value{
density of target distribution in MCMC Sampler of the proposal
distribution for the coefficients
}
\description{
See \code{density_wald}.
}
