% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_lee.R
\name{true_chen_lee}
\alias{true_chen_lee}
\title{Obtain true coefficients on endogeneous variable}
\usage{
true_chen_lee(beta_D_errors, tau)
}
\arguments{
\item{beta_D_errors}{coefficients on the interaction of endogeneous
variables and errors}

\item{tau}{quantile between 0 and 1}
}
\value{
Vector of true DGP coefficients on endogeneous variables
}
\description{
See third display on p. 557 of Chen and Lee (2018)
}
\seealso{
\code{\link{chen_lee}}
}
