% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald_varcov}
\alias{wald_varcov}
\title{Compute Wald-type asymptotic variance}
\usage{
wald_varcov(
  resid,
  alpha,
  tau,
  D,
  X,
  Z,
  Phi = linear_projection(D, X, Z),
  psi = 1
)
}
\arguments{
\item{resid}{MILP residuals associated with point estimate}

\item{alpha}{Alpha level of the test; defaults to 0.1; used to compute
critical value and the Hall and Sheather bandwidth (for estimating
J_{Psi}(tau) in CH (2006))}

\item{tau}{Quantile (number between 0 and 1)}

\item{D}{Matrix of endogeneous variables}

\item{X}{Matrix of covariates (including intercept)}

\item{Z}{Matrix of instruments (only relevant for obtaining \code{Phi})}

\item{Phi}{Transformed matrix of instruments; defaults to linear projection
of D on X and Z}

\item{psi}{Scalar coefficient in front of the variance-covariance matrix;
useful for tuning MCMC subsampling (see \code{mcmc_active_basis}); defaults
to 1}
}
\value{
Named list
\enumerate{
\item \code{varcov}: variance-covariance matrix multiplied by \code{psi}
\item \code{hs}: Hall and Sheather Bandwidth
\item Other objects used to create varcov
}
}
\description{
See (3.11) from Chernozhukov and Hansen (2006)
}
