% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-h.R
\name{weight_indices}
\alias{weight_indices}
\title{Compute weights on indices to propose active basis}
\usage{
weight_indices(residuals_opt, theta)
}
\arguments{
\item{residuals_opt}{Residuals from IQR-QR problem}

\item{theta}{Tuning parameter}
}
\value{
weight for each index
}
\description{
Indices whose residuals are small are weighted more than indices whose
residuals are large.
}
