% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{write_prm}
\alias{write_prm}
\title{Save Gurobi Parameters}
\usage{
write_prm(params, path, filename)
}
\arguments{
\item{params}{A list of Gurobi parameters}

\item{path}{The directory to store the resulting \code{.prm} file}

\item{filename}{The name of the resulting \code{.prm} file}
}
\value{
A \code{.prm} file named \code{filename} in the directory \code{path}
that stores \code{params}
}
\description{
Create a .prm file with parameters from Gurobi
}
\details{
While \code{gurobi::gurobi_write()} can save Gurobi models, it cannot save
the parameters. The function \code{write_prm} overcomes this limitation
by creating a \code{.prm} file with the parameters.
}
\examples{
\dontrun{
params <- list(NonConvex = 0)
write_prm(params, ".", "example")
file.remove("./example.prm")
}

}
\seealso{
\code{\link[gurobi]{gurobi_write}}
}
\concept{debugging}
