% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch-aux-fcns.R
\name{get_initial_beta_D}
\alias{get_initial_beta_D}
\title{Obtain naive coefficient estimates for \code{beta_D}}
\usage{
get_initial_beta_D(Y, D, Z, X, ...)
}
\arguments{
\item{Y}{Outcome matrix (more precisely, a column vector)}

\item{D}{Matrix of endogenous variables}

\item{Z}{Matrix of instruments}

\item{X}{Matrix of covariates (incl. column of 1's for the intercept)}

\item{...}{Arguments to be passed to \code{quantreg::rq()}}
}
\value{
A vector of coefficients on the endogenous variable
}
\description{
Naively regress \code{Y} on \code{D}, \code{Z}, and \code{X}, and return
the coefficients on \code{D}.
}
