% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_milp.R
\name{iqr_milp}
\alias{iqr_milp}
\title{Compute inverse quantile regression estimator}
\usage{
iqr_milp(
  Y,
  X,
  D,
  Z,
  tau,
  O_neg,
  O_pos,
  M,
  params = list(TimeLimit = 300, FeasibilityTol = 1e-06, OutputFlag = 0),
  quietly = TRUE
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile (number between 0 and 1)}

\item{O_neg, O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 10}

\item{params}{Gurobi parameters, see \url{https://www.gurobi.com/documentation/9.1/refman/parameter_descriptions.html}}

\item{quietly}{If TRUE (default), sends messages during execution (boolean)}
}
\description{
Solve a mixed integer linear program to compute the inverse
quantile regression estimator.
}
