% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_milp.R
\name{iqr_milp}
\alias{iqr_milp}
\title{Compute inverse quantile regression estimator}
\usage{
iqr_milp(
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  O_neg = NULL,
  O_pos = NULL,
  M = NULL,
  TimeLimit = 300,
  VarHintVal_bool = FALSE,
  VarHintPri_bool = FALSE,
  BranchPriority_bool = FALSE,
  sparse = TRUE,
  attributes = list(),
  params = list(FeasibilityTol = 1e-06, LogToConsole = 0),
  start_method = NULL,
  start = NULL,
  fix = NULL,
  quietly = TRUE,
  show_progress = TRUE,
  LogFileName = "",
  LogFileExt = ".log"
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{O_neg, O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 2 times the largest absolute residual from
quantile regression of Y on X and D}

\item{TimeLimit}{Maximum time (in seconds) spent on a linear program;
defaults to 300, will be appended to \code{params}}

\item{VarHintVal_bool}{If TRUE, instead of fixing the binary variables as
determined by \code{O_neg,O_pos}, we will "hint" these values; defaults to
FALSE; see
\url{https://www.gurobi.com/documentation/9.1/refman/varhintval.html#attr:VarHintVal};
overrides \code{attributes}}

\item{VarHintPri_bool}{If TRUE, we specify the strength of the "hints"
according to the magnitude of the residual from a naive QR;
only valid if VarHintVal_bool is TRUE; overrides \code{attributes};
defaults to FALSE;
see \url{https://www.gurobi.com/documentation/9.1/refman/varhintpri.html#attr:VarHintPri}}

\item{BranchPriority_bool}{If TRUE, we specify which variables will be
branched on first according to the magnitude of the residual from a naive QR;
defaults to FALSE;
see \url{https://www.gurobi.com/documentation/9.1/refman/branchpriority.html}}

\item{sparse}{If TRUE (default), use sparse matrices}

\item{attributes}{named list of Gurobi attributes, see
\url{https://www.gurobi.com/documentation/9.1/refman/attributes.html} and
\url{https://www.gurobi.com/documentation/9.1/refman/the_model_argument.html};
defaults to empty list}

\item{params}{Gurobi parameters, see
\url{https://www.gurobi.com/documentation/9.1/refman/parameter_descriptions.html}}

\item{start_method}{Feed into \code{compute_warmstart} in the default value
of \code{start}; defaults to NULL, ignored if \code{start} is not NULL}

\item{start}{Gurobi attribute, see
\url{https://www.gurobi.com/documentation/9.1/examples/mip_starts.html}; If
NULL (default), no starting solution will be provided; see \code{compute_warmstart}}

\item{fix}{Fix decision variables; If NULL (default), no starting solution will
be provided; if NA, the respective decision variable won't be fixed;
decision variables are:
\enumerate{
\item beta_X
\item beta_Phi_plus
\item beta_Phi_minus ; note: abs(beta_Phi) = beta_Phi_plus + beta_Phi_minus
\item beta_D
\item u
\item v
\item a
\item k
\item l
}}

\item{quietly}{If TRUE (default), supress messages during execution (boolean)}

\item{show_progress}{If TRUE (default), sends progress messages during
execution (boolean)}

\item{LogFileName}{Name of Gurobi log file; If string is empty (default),
Gurobi log won't be saved (string)}

\item{LogFileExt}{Extension of Gurobi log file; If \code{LogFileName} is
empty, then Gurobi log won't be saved and this argument will be ignored;
defaults to "log" (string)}
}
\value{
A named list of
\enumerate{
\item iqr: Gurobi model that was solved
\item params: Gurobi parameters used
\item result: solution to MILP returned by Gurobi
\item status: status of Gurobi's solution
\item beta_X: coefficients on exogenous variables
\item beta_Phi_plus: positive part of coefficients on instruments
\item beta_Phi_minus: negative part of coefficients on instruments
\item beta_D: coefficients on endogenous variables
\item u: positive part of residuals
\item v: negative part of residuals
\item a: dual variable
\item k: binary variable associated with u
\item l: binary variable associated with v
\item beta_Phi: coefficients on instruments
(beta_Phi_plus - beta_Phi_minus)
\item resid: residuals (u - v)
\item objval: value of objective function (absolute value of beta_Phi)
}
}
\description{
Solve a mixed integer linear program to compute the inverse
quantile regression estimator.
}
