% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_milp.R
\name{iqr_milp}
\alias{iqr_milp}
\title{Compute inverse quantile regression estimator}
\usage{
iqr_milp(
  Y,
  X,
  D,
  Z,
  tau,
  O_neg = NULL,
  O_pos = NULL,
  M = NULL,
  TimeLimit = 300,
  projection = TRUE,
  params = list(FeasibilityTol = 1e-06, OutputFlag = 0),
  quietly = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{tau}{Quantile (number between 0 and 1)}

\item{O_neg, O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 10 * standard deviation of residuals from
quantile regression of Y on X and D}

\item{TimeLimit}{Maximum time (in seconds) spent on a linear program;
defaults to 300, will be appended to \code{params}}

\item{projection}{If TRUE (default), project D on the space spanned by X and
Z to construct the vector of functions of transformed instruments; else,
let Z be the instruments for endogenous variables (boolean)}

\item{params}{Gurobi parameters, see \url{https://www.gurobi.com/documentation/9.1/refman/parameter_descriptions.html}}

\item{quietly}{If TRUE (default), sends messages during execution (boolean)}

\item{show_progress}{If TRUE (default), sends progress messages during execution (boolean)}
}
\value{
A named list of
\enumerate{
\item iqr: Gurobi model that was solved
\item params: Gurobi parameters used
\item result: solution to MILP returned by Gurobi
\item status: status of Gurobi's solution
\item beta_X: coefficients on exogenous variables
\item beta_Phi_plus: positive part of coefficients on instruments
\item beta_Phi_minus: negative part of coefficients on instruments
\item beta_D: coefficients on endogenous variables
\item u: positive part of residuals
\item v: negative part of residuals
\item a: dual variable
\item k: binary variable associated with u
\item l: binary variable associated with v
\item beta_Phi: coefficients on instruments (beta_Phi_plus - beta_Phi_minus)
\item resid: residuals (u - v)
\item objval: value of objective function (absolute value of beta_Phi)
}
}
\description{
Solve a mixed integer linear program to compute the inverse
quantile regression estimator.
}
