% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivqr.R
\name{ivqr}
\alias{ivqr}
\title{IVQR: Instrumental Variable Quantile Regression}
\usage{
ivqr(y, e, X, Z, tau = 0.5, method = "br", grilla = seq(-10, 10, 0.1))
}
\arguments{
\item{y}{Vector de la variable dependiente.}

\item{e}{Vector de la variable endogena.}

\item{X}{Matriz de variables explicativas.}

\item{Z}{Matriz de variables instrumentales.}

\item{tau}{Cuantil a estimar (por defecto, la mediana).}

\item{method}{Método de optimización (por defecto "br" para Barrodale-Roberts).}

\item{grilla}{Grilla de valores para la busqueda del coeficiente de e.}
}
\value{
Una lista con los coeficientes, residuales y valores ajustados.
}
\description{
Esta función implementa el modelo de regresión cuantil con variables instrumentales (IVQR) siguiendo el enfoque de Chernozhukov y Hansen.
}
