% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqcp_proj.R
\name{miqcp_proj}
\alias{miqcp_proj}
\title{Subvector inference under weak identification with MIQCP}
\usage{
miqcp_proj(
  j,
  alpha = 0.1,
  sense,
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  orthogonalize_statistic = FALSE,
  O_neg = NULL,
  O_pos = NULL,
  M = NULL,
  TimeLimit = 300,
  params = list(FeasibilityTol = 1e-06, OutputFlag = 0),
  quietly = TRUE,
  show_progress = TRUE,
  LogFileName = "",
  LogFileExt = ".log"
)
}
\arguments{
\item{j}{Index of D associated with the coefficient of interest
(numeric between 1 and p_D)}

\item{alpha}{Alpha level; defaults to 10\% (numeric between 0 and 1)}

\item{sense}{Maximize or minimize beta_D,j (either "max" or "min")}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{orthogonalize_statistic}{If TRUE, \eqn{\tilde{B}} will be used in
numerator of test statistic; defaults to FALSE; for advanced users only}

\item{O_neg}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 10 * standard deviation of residuals from
quantile regression of Y on X and D}

\item{TimeLimit}{Maximum time (in seconds) spent on a linear program;
defaults to 300, will be appended to \code{params}}

\item{params}{Gurobi parameters, see
\url{https://www.gurobi.com/documentation/9.1/refman/parameter_descriptions.html}}

\item{quietly}{If TRUE (default), sends messages during execution (boolean)}

\item{show_progress}{If TRUE (default), sends progress messages during
execution (boolean)}

\item{LogFileName}{Name of Gurobi log file; If string is empty (default),
Gurobi log won't be saved (string)}

\item{LogFileExt}{Extension of Gurobi log file; If \code{LogFileName} is
empty, then Gurobi log won't be saved and this argument will be ignored;
defaults to "log" (string)}
}
\value{
A named list of
\enumerate{
\item proj: Gurobi model that was solved
\item params: Gurobi parameters used
\item result: solution to MILP returned by Gurobi
\item status: status of Gurobi's solution
\item beta_X: coefficients on exogenous variables
\item beta_D: coefficients on endogenous variables
\item u: positive part of residuals
\item v: negative part of residuals
\item a: dual variable
\item k: binary variable associated with u
\item l: binary variable associated with v
\item resid: residuals (u - v)
\item objval: value of objective function (beta_D,j)
}
}
\description{
Invert the null on the full weakly identified vector of endogeneous
coefficients and maximize/minimize the projection of the resulting
multivariate confidence on the axis of a particular coefficient.
}
\details{
The maximization and minimization of the projection is formulated as a
Mixed Integer Quadratically Constrained Program (MIQCP).
}
