% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqcp_proj.R
\name{miqcp_proj}
\alias{miqcp_proj}
\title{Subvector inference under weak identification with MIQCP}
\usage{
miqcp_proj(
  projection_index,
  endogeneous = TRUE,
  beta_X_indices = NULL,
  alpha = 0.1,
  sense,
  Y,
  X,
  D,
  Z,
  Phi = linear_projection(D, X, Z),
  tau,
  orthogonalize_statistic = FALSE,
  homoskedasticity = FALSE,
  kernel = "Powell",
  residuals = NULL,
  O_neg = NULL,
  O_pos = NULL,
  M = NULL,
  TimeLimit = 300,
  params = list(FeasibilityTol = 1e-06, OutputFlag = 0),
  sparse = TRUE,
  quietly = TRUE,
  show_progress = TRUE,
  LogFileName = "",
  LogFileExt = ".log"
)
}
\arguments{
\item{projection_index}{Index associated with the coefficient of interest
(numeric between 1 and p_D if \code{endogeneous} is TRUE;
numeric between 1 and p_X if \code{endogeneous} is FALSE)}

\item{endogeneous}{If TRUE (default), the function will project on the axis
of the corresponding endogeneous variable's coefficient; if FALSE, the
function will project on the axis of the corresponding exogeneous variable's
coefficient (boolean)}

\item{beta_X_indices}{Indices of endogeneous variables to be included in the
null hypothesis; if NULL (default), none of the coefficients on the
endogeneous variables will be specified in the null (numeric vector)}

\item{alpha}{Alpha level; defaults to 10\% (numeric between 0 and 1)}

\item{sense}{Maximize or minimize beta_D,j (either "max" or "min")}

\item{Y}{Dependent variable (vector of length n)}

\item{X}{Exogenous variable (including constant vector) (n by p_X matrix)}

\item{D}{Endogenous variable (n by p_D matrix)}

\item{Z}{Instrumental variable (n by p_Z matrix)}

\item{Phi}{Transformation of X and Z to be used in the program;
defaults to the linear projection of D on X and Z (matrix with n rows)}

\item{tau}{Quantile (number between 0 and 1)}

\item{orthogonalize_statistic}{If TRUE, \eqn{\tilde{B}} will be used in
numerator of test statistic; defaults to FALSE; for advanced users only}

\item{homoskedasticity}{If TRUE, assume density of error at 0 is constant;
defaults to FALSE (boolean)}

\item{kernel}{Only active if \code{homoskedasticity} is FALSE; either
"Powell" (default) to use the Powell estimator or
"Gaussian" to use a Gaussian kernel; only used when
\code{homoskedasticity} is FALSE}

\item{residuals}{Residuals from IQR MILP program; if NULL (default), use
naive residuals from quantile regression}

\item{O_neg, O_pos}{Indices for residuals whose sign is fixed to be negative
and positive, respectively (vectors)}

\item{M}{A large number that bounds the absolute value of the residuals
(a positive number); defaults to 2 times the largest absolute residual from
quantile regression of Y on X and D}

\item{TimeLimit}{Maximum time (in seconds) spent on a linear program;
defaults to 300, will be appended to \code{params}}

\item{params}{Gurobi parameters, see
\url{https://www.gurobi.com/documentation/9.1/refman/parameter_descriptions.html}}

\item{sparse}{If TRUE (default), use sparse matrices}

\item{quietly}{If TRUE (default), supress messages during execution (boolean)}

\item{show_progress}{If TRUE (default), sends progress messages during
execution (boolean)}

\item{LogFileName}{Name of Gurobi log file; If string is empty (default),
Gurobi log won't be saved (string)}

\item{LogFileExt}{Extension of Gurobi log file; If \code{LogFileName} is
empty, then Gurobi log won't be saved and this argument will be ignored;
defaults to "log" (string)}
}
\value{
A named list of # TODO: update
\enumerate{
\item proj: Gurobi model that was solved
\item params: Gurobi parameters used
\item result: solution to MILP returned by Gurobi
\item status: status of Gurobi's solution
\item beta_X: coefficients on exogenous variables
\item beta_D: coefficients on endogenous variables
\item u: positive part of residuals
\item v: negative part of residuals
\item a: dual variable
\item k: binary variable associated with u
\item l: binary variable associated with v
\item resid: residuals (u - v)
\item objval: value of objective function (beta_D,j)
\item M: big M constant used for complementary slackness conditions
\item Phi_J,Phi_J_minus,X_K,X_K_minus,B,B_tilde,Psi: matrices used in
program
\item projection_index,endogeneous: coefficient onto which the
multivariate confidence region was projected
\item homoskedasticity,kernel: indicates estimator of residual density
}
}
\description{
Invert the null on the full weakly identified vector of endogeneous
coefficients and project the resulting multivariate confidence region on the
axis of a specified coefficient.
}
\details{
The maximization and minimization of the projection is formulated as a
Mixed Integer Quadratically Constrained Program (MIQCP).

The index of projection is given by \code{projection_index} and
\code{endogeneous}.  For example, if \code{projection_index} is 1 and
\code{endogeneous} is TRUE, we project on the axis of \eqn{\beta_{D, 1}}.
If \code{projection_index} is 2 and \code{endogeneous} is FALSE, we project
on the axis of \eqn{\beta_{X, 2}}.

Under weak identification, we invert the null hypothesis on the full weakly
identified vector of endogeneous coefficients.  To include exogeneous
coefficients in the null, specify the desired indices of the exogeneous
variables in \code{beta_X_indices}.
}
