% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald_univariate}
\alias{wald_univariate}
\title{Compute Wald-type univariate confidence interval}
\usage{
wald_univariate(
  center,
  endogeneous,
  index,
  resid,
  alpha = 0.1,
  tau,
  D,
  X,
  Z,
  Phi = linear_projection(D, X, Z)
)
}
\arguments{
\item{center}{MILP point estimate that is the center of the univariate
confidence interval (scalar)}

\item{endogeneous}{If TRUE (default), \code{index} refers to the
coefficients on the endogeneous variables; if FALSE, \code{index} refers to
the coefficients on the exogeneous variables (boolean)}

\item{index}{Index of the coefficient of interest
(numeric between 1 and p_D if \code{endogeneous} is TRUE;
numeric between 1 and p_X if \code{endogeneous} is FALSE)}

\item{resid}{MILP residuals associated with point estimate}

\item{alpha}{Alpha level of the test; defaults to 0.1; used to compute
critical value and the Hall and Sheather bandwidth (for estimating
J_{Psi}(tau) in CH (2006))}

\item{tau}{Quantile (number between 0 and 1)}

\item{D}{Matrix of endogeneous variables}

\item{X}{Matrix of covariates (including intercept)}

\item{Z}{Matrix of instruments (only relevant for obtaining \code{Phi})}

\item{Phi}{Transformed matrix of instruments; defaults to linear projection
of D on X and Z}
}
\description{
See (3.11) from Chernozhukov and Hansen (2006).
}
