% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVQR.R
\name{weakIVtest}
\alias{weakIVtest}
\title{Weak-IV robust test}
\usage{
weakIVtest(object, size = 0.05)
}
\arguments{
\item{object}{An ivqr object returned from the function \code{ivqr()}}

\item{size}{A number indicating the size of the test. Default is 0.05.
package quantreg.}
}
\value{
An ivqr_weakIV object, which will be directly inputted into the method
\code{plot.ivqr_weakIV()} to visualize the confidence 
region. It is also possible to print the confidence region by applying the method 
\code{print()} to the ivqr_weakIV object. 
Note that the confidence region is not guaranteed to be convex.
}
\description{
Weak-IV robust test
}
\examples{
data(ivqr_eg)
fit <- ivqr(y ~ d | z | x, seq(0.1,0.9,0.1), grid = seq(-2,2,0.2), data = ivqr_eg) # a process
weakIVtest(fit)
print(fit) # print the confidence region.
}
